Fix typing error in sqlx queries

+++ src/dto.rs
@@ -1,8 +1,8 @@
 #[derive(sqlx::FromRow)]
 pub struct SubscriptionDto {
-    pub id: i32,
-    pub chat_id: i32,
-    pub is_enabled: i32,
+    pub id: i64,
+    pub chat_id: i64,
+    pub is_enabled: i64,
     pub created_at: i64,
     pub updated_at: i64,
 }
